requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
tiktoken==0.4.0
huggingface-hub==0.14.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2

altair	==			4.2.0
attrs	==			23.1.0
bcrypt	==			4.0.1
blinker	==			1.6.2
cachetools	==		5.3.1
certifi	==			2023.7.22
charset-normalizer	==	3.2.0
click	==			8.1.7
colorama	==		0.4.6
contourpy	==		1.1.1
cycler	==			0.11.0
dacite	==			1.8.1
entrypoints	==		0.4
et-xmlfile	==		1.1.0
extra-streamlit-components ==	0.1.56
fonttools	==		4.42.1
gitdb	==			4.0.10
GitPython	==		3.1.35
htmlmin	==			0.1.12
idna	==			3.4
ImageHash	==		4.3.1
importlib-metadata	==	6.8.0
importlib-resources	==	6.1.0
Jinja2	==			3.1.2
joblib	==			1.3.2
jsonschema	==		4.19.0
jsonschema-specifications==	2023.7.1
kaleido	==			0.2.1
kiwisolver	==		1.4.5
markdown-it-py	==		3.0.0
MarkupSafe	==		2.1.3
matplotlib	==		3.7.0
mdurl	==			0.1.2
networkx	==		3.1
numerize	==		0.12
numpy	==			1.23.5
openpyxl	==		3.0.10
packaging	==		23.1
pandas	==			1.5.2
pandas-profiling	==	3.6.6
patsy	==			0.5.3
phik	==			0.12.3
Pillow	==			10.0.0
pip	==			23.2.1
plotly	==			5.11.0
protobuf	==		3.20.3
pyarrow	==			13.0.0
pydantic	==		1.10.13
pydeck	==			0.8.1b0
Pygments	==		2.16.1
PyJWT	==			2.8.0
Pympler	==			1.0.1
pyparsing	==		3.1.1
python-dateutil	==		2.8.2
python-decouple	==		3.8
pytz	==			2023.3.post1
PyWavelets	==		1.4.1
PyYAML	==			6.0.1
referencing	==		0.30.2
requests	==		2.31.0
rich	==			13.5.2
rpds-py	==			0.10.2
scikit-learn	==		1.1.3
scipy	==			1.9.3
seaborn	==			0.12.2
semver	==			3.0.1
setuptools	==		68.1.2
six	==			1.16.0
smmap	==			5.0.0
statsmodels	==		0.14.0
streamlit-aggrid	==	0.3.4.post3
streamlit-authenticator	==	0.2.1
streamlit-pandas-profiling==	0.1.3
sweetviz	==		2.2.1
tangled-up-in-unicode	==	0.2.0
tenacity	==		8.2.3
threadpoolctl	==		3.2.0
toml	==			0.10.2
toolz	==			0.12.0
tornado	==			6.3.3
tqdm	==			4.66.1
typeguard	==		2.13.3
typing_extensions	==	4.7.1
tzdata	==			2023.3
tzlocal	==			5.0.1
urllib3	==			2.0.4
validators	==		0.22.0
visions	==			0.7.5
watchdog	==		3.0.0
wheel	==			0.41.2
wordcloud	==		1.9.2
ydata-profiling	==		4.5.1
zipp	==			3.16.2
